separate logic
